20_04_Proj2_Mongo_conex

+++ src/controllers/games.controller.js
@@ -7,7 +7,17 @@ const findGamesController = (req, res) => {
 
 const findGameByIdController = (req, res) => {
   const idParam = req.params.id;
+
+  if (!idParam) {
+    return res.status(400).send({ message: "ID não informado!" })
+  }
+
   const chosenGame = gamesService.findGameByIdService(idParam);
+
+  if (!chosenGame) {
+    return res.status(404).send({ message: "Game não encontrado!" })
+  }
+
   res.send(chosenGame);
 };
 
@@ -30,16 +40,31 @@ const createGameController = (req, res) => {
   const updateGameController = (req, res) => {
     const idParam = +req.params.id;
     const gameEdit = req.body;
+
+    if (!idParam) {
+      return res.status(404).send({ message: "Game não encontrado!" })
+    }
+  
+    if (!gameEdit || !gameEdit.nome || !gameEdit.empresa || !gameEdit.ano) {
+      return res.status(400).send({ message: "Você não preencheu todos os dados para editar o game!" });
+    }
+
     const updatedGame = gamesService.updateGameService(idParam, gameEdit);
     res.send(updatedGame);
   };
   
   const deleteGameController = (req, res) => {
     const idParam = req.params.id;
+
+    if (!idParam) {
+      return res.status(404).send({ message: "Game não encontrado!" })
+    }
+
     gamesService.deleteGameService(idParam);
     res.send({ message: 'Game deletado com sucesso!' });
   };
 
+  
 module.exports = {
   findGamesController,
   findGameByIdController,
